+++ app/src/main/java/com/solux/flory/presentation/profile/ProfileFragment.kt
@@ -4,12 +4,14 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import androidx.navigation.fragment.findNavController
+import com.solux.flory.R
 import com.solux.flory.databinding.FragmentProfileBinding
+import com.solux.flory.presentation.auth.LoginActivity
 import com.solux.flory.util.base.BindingFragment
 
 class ProfileFragment : BindingFragment<FragmentProfileBinding>(FragmentProfileBinding::inflate) {
     private val viewModel by activityViewModels<ProfileViewModel>()
-    private lateinit var adapter: ProfileAdapter
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initAdapter()
@@ -19,15 +21,15 @@ class ProfileFragment : BindingFragment<FragmentProfileBinding>(FragmentProfileB
     }
 
     private fun initAdapter() {
-        adapter = ProfileAdapter()
-        binding.rvProfileNeighbors.adapter = adapter
-        adapter.submitList(viewModel.mockNeighbors)
+        ProfileAdapter().apply {
+            binding.rvProfileNeighbors.adapter = this
+            submitList(viewModel.mockNeighbors)
+        }
     }
 
     private fun profileModifyTextClick() {
         binding.tvProfileModification.setOnClickListener {
-            val intent = Intent(requireContext(), ProfileModifyActivity::class.java)
-            startActivity(intent)
+            findNavController().navigate(R.id.action_fragment_profile_to_fragment_profile_modify)
         }
     }
 
@@ -40,7 +42,10 @@ class ProfileFragment : BindingFragment<FragmentProfileBinding>(FragmentProfileB
 
     private fun logoutBtnClick() {
         binding.btnProfileLogout.setOnClickListener {
-            // 커스텀 다이얼로그 띄우기
+            Intent(requireContext(), LoginActivity::class.java).apply {
+                addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
+                startActivity(this)
+            }
         }
     }
 }
