+++ app/src/main/res/layout/fragment_profile_modify.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.profile.ProfileModifyActivity">
+    tools:context=".presentation.profile.ProfileModifyFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_modify_header"
